requirements.txt
boto3
click
PyYAML
beeprint
